+++ CHANGES.rst
@@ -2,6 +2,7 @@
 Revision 0.2.2, released XX-07-2018
 -----------------------------------
 
+- Added Apache expression tags (ap_expr) test case
 - Added support for quoting named blocks
 - Added `includeoptinal` apacheinclude statement support
 - Include and Apache include statements made case-insensitive